requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coverage==6.4.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.0.0
zipp==3.8.1
